pyproject.toml
[project]
name = "mcp-server-search"
version = "0.6.2"
description = "An MCP (Model Context Protocol) server that provides Google search functionality for AI models"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Anthropic, PBC." }]
maintainers = [{ name = "Maarten Smeets", email = "wisp1980@gmail.com" }]
keywords = ["http", "mcp", "llm", "automation"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "googlesearch-python==1.3.0",
    "fake-useragent==1.4.0",
    "portalocker==2.8.2",
    "pydantic>=2.0.0",
    "fastmcp==2.8.0"
]

[project.scripts]
mcp-server-search = "mcp_server_search:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "ruff>=0.7.3"]
